requirements.txt
absl-py==2.2.2
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
biopython==1.85
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1739515848642/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy==1.8.14
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1733569351617/work
fonttools @ file:///C:/b/abs_4crkswws2h/croot/fonttools_1737040078745/work
grpcio==1.71.0
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1738273805233/work
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython @ file:///D:/bld/ipython_1701831845989/work
ipywidgets @ file:///home/conda/feedstock_root/build_artifacts/ipywidgets_1744294556535/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_widgets_1744291043727/work
kiwisolver @ file:///C:/b/abs_88mdhvtahm/croot/kiwisolver_1672387921783/work
Markdown==3.8
MarkupSafe==3.0.2
matplotlib @ file:///C:/b/abs_ae02atcfur/croot/matplotlib-suite_1667356722968/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
nest-asyncio==1.6.0
numpy @ file:///C:/Users/dev-admin/mkl/numpy_and_numpy_base_1682982345978/work
packaging @ file:///C:/b/abs_3by6s2fa66/croot/packaging_1734472138782/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow @ file:///C:/b/abs_68t826txdy/croot/pillow_1744613085333/work
platformdirs==4.3.7
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1737453357274/work
protobuf==6.30.2
psutil==7.0.0
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
py3Dmol==2.4.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
pyparsing @ file:///C:/b/abs_40z8gyj9wi/croot/pyparsing_1731445739241/work
PyQt6==6.7.1
PyQt6_sip @ file:///C:/b/abs_28s7k4h_hl/croot/pyqt-split_1740498234166/work/pyqt_sip
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil @ file:///C:/b/abs_3au_koqnbs/croot/python-dateutil_1716495777160/work
pywin32==310
pyzmq==26.4.0
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
sip @ file:///C:/b/abs_5cto136kse/croot/sip_1738856220313/work
six @ file:///C:/b/abs_149wuyuo1o/croot/six_1744271521515/work
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tomli @ file:///C:/Windows/TEMP/abs_ac109f85-a7b3-4b4d-bcfd-52622eceddf0hy332ojo/croots/recipe/tomli_1657175513137/work
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado @ file:///C:/b/abs_7cyu943ybx/croot/tornado_1733960510898/work
tqdm @ file:///C:/ci/tqdm_1650636210717/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
unicodedata2 @ file:///C:/b/abs_dfnftvxi4k/croot/unicodedata2_1736543771112/work
urllib3 @ file:///C:/b/abs_7bst06lizn/croot/urllib3_1737133657081/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
Werkzeug==3.1.3
widgetsnbextension @ file:///home/conda/feedstock_root/build_artifacts/widgetsnbextension_1744291053367/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
zipp==3.21.0
